CSS Retouch + Delete All Games

See merge request polytechnique-montr-al/log2990/20231/equipe-304/LOG2990-304!152

+++ client/src/app/pages/configuration-page/configuration-page.component.ts
@@ -1,6 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { Dialogs } from '@app/interfaces/dialogs';
 import { LevelService } from '@app/services/level/level.service';
+import { PopUpService } from '@app/services/pop-up/pop-up.service';
 
 /**
  * This component represents the page where the user can modify, delete or create new games. It is accessible from
@@ -14,11 +16,16 @@ import { LevelService } from '@app/services/level/level.service';
     templateUrl: './configuration-page.component.html',
     styleUrls: ['./configuration-page.component.scss'],
 })
-export class ConfigurationPageComponent implements OnInit {
-    constructor(public levelService: LevelService, public router: Router) {}
+export class ConfigurationPageComponent implements OnInit, OnDestroy {
+    constructor(public levelService: LevelService, private popUpService: PopUpService, public router: Router) {}
 
     ngOnInit(): void {
         this.levelService.refreshLevels();
+        this.levelService.setupSocket();
+    }
+
+    ngOnDestroy(): void {
+        this.levelService.destroySocket();
     }
 
     /**
@@ -28,6 +35,42 @@ export class ConfigurationPageComponent implements OnInit {
      * @param levelId the id of the level to delete.
      */
     onDeleteLevel(levelId: number): void {
-        this.levelService.deleteLevel(levelId);
+        this.popUpService.openDialog(Dialogs.confirmDeleteLevel);
+        this.popUpService.dialogRef.afterClosed().subscribe((confirmation) => {
+            if (confirmation) {
+                this.levelService.deleteLevel(levelId);
+            }
+        });
+    }
+
+    /**
+     * Event listener for the delete all levels button.
+     * It will call the deleteAllLevels function from the levelService.
+     */
+    onDeleteAllLevels(): void {
+        this.popUpService.openDialog(Dialogs.confirmDeleteAllLevels);
+        this.popUpService.dialogRef.afterClosed().subscribe((confirmation) => {
+            if (confirmation) {
+                this.levelService.deleteAllLevels();
+            }
+        });
+    }
+
+    onResetLevelHighScore(levelId: number): void {
+        this.popUpService.openDialog(Dialogs.confirmResetLevelHighScore);
+        this.popUpService.dialogRef.afterClosed().subscribe((confirmation) => {
+            if (confirmation) {
+                this.levelService.resetLevelHighScore(levelId);
+            }
+        });
+    }
+
+    onResetGameConstants(): void {
+        this.popUpService.openDialog(Dialogs.confirmResetGameConstants);
+        this.popUpService.dialogRef.afterClosed().subscribe((confirmation) => {
+            if (confirmation) {
+                this.levelService.resetGameConstants();
+            }
+        });
     }
 }
